+++ src/blocks/accordion/deprecation.tsx
@@ -8,7 +8,7 @@ import type {
 	CreateBlockSaveProps,
 } from "@atomicsmash/blocks-helpers";
 import { useBlockProps, InnerBlocks, RichText } from "@wordpress/block-editor";
-import { Icon } from "@plugin/blocks/svgs";
+import { Icon } from "@launchpadBlocks/svgs";
 import { attributes as currentAttributes } from "./attributes";
 import { supports as currentSupports } from "./supports";