+++ src/blocks/carousel-thumbnails/index.tsx
@@ -2,7 +2,7 @@ import type { InterpretedAttributes } from "./attributes";
 import type { InterpretedUsedContext } from "./context";
 import { registerBlockType } from "@atomicsmash/blocks-helpers";
 import { registerBlockVariation } from "@wordpress/blocks";
-import { registerLaunchpadBlocksCollection } from "@plugin/blocks/helpers.editor";
+import { registerLaunchpadBlocksCollection } from "@launchpadBlocks/helpers.editor";
 // import { Icon } from "@blocks/svgs";
 import blockMetaData from "./block.json";
 import { deprecated } from "./deprecation";