+++ src/blocks/fancy-list-item/index.tsx
@@ -1,7 +1,7 @@
 import type { InterpretedAttributes } from "./attributes";
 import type { InterpretedUsedContext } from "./context";
 import { registerBlockType } from "@atomicsmash/blocks-helpers";
-import { registerLaunchpadBlocksCollection } from "@plugin/blocks/helpers.editor";
+import { registerLaunchpadBlocksCollection } from "@launchpadBlocks/helpers.editor";
 // import { Icon } from "@blocks/svgs";
 import blockMetaData from "./block.json";
 import { Edit } from "./edit";