+++ src/blocks/nav-list/index.tsx
@@ -1,7 +1,7 @@
 import type { InterpretedAttributes } from "./attributes";
 import { registerBlockType } from "@atomicsmash/blocks-helpers";
 import { navigation } from "@wordpress/icons";
-import { registerLaunchpadBlocksCollection } from "@plugin/blocks/helpers.editor";
+import { registerLaunchpadBlocksCollection } from "@launchpadBlocks/helpers.editor";
 import blockMetaData from "./block.json";
 import { Edit } from "./edit";
 import { save } from "./save";